backend/requirements.txt
fastapi
uvicorn
google-api-python-client
pymongo
python-dotenv
textblob
pandas
matplotlib
seaborn